# speed up filter of vector