@chethega 2019-05-06
I am not entirely sure whether we can give `filter!(f, a::AbstractVector)` the same treatment.

For this, we would do
```
function filter!(f, a::AbstractVector)
    j = firstindex(a)
    for ai in a
        @inbounds a[j] = ai
        j = ifelse(f(ai), j+1, j)
    end
    deleteat!(a, j:lastindex(a))
    sizehint!(a, length(a))
    a
end
```
This would work for simple offset vectors. Are there valid abstract vectors where this would fail?

Are we OK with returning a `Vector` on `filter(pred, a::AbstractVector)`? If not, what is the right incantation to create an `undef` variant of `a` (same offsets, indices, type)?