@chethega 2019-05-07
I think this could be ready to go: All dispatches now rely on the same interface functions as before (`getindex` for out-of-place filtering of abstract vectors, `deleteat!` for inplace filtering of abstract vectors, unchanged code for higher dimensional arrays, known existing interface functions for built-in vectors). I'll squash the commits when ready.

The test failures look unrelated to me. Can we re-run the tests? Can we get a nanosoldier to catch unexpected perf regressions and otherwise delight in smaller numbers?


Then there is the question of time-memory tradeoff: The `realloc` after filtering is kinda expensive. With the expensive realloc (currently conservatively included), this PR is a strict improvement in both space and time, unless the branches are predicted perfectly (e.g. take every second element). 

Any votes for speeding this up even further, at the cost of potentially increased memory consumption? I tend to be against: The overhead of `filter` is now a small factor of `copy`, which should be fast enough in case of cheap predicates and is irrelevant in case of expensive predicates.