@KristofferC 2019-05-07
All CI seems to fail with the same tests, are you sure they are unrelated? I would expect they are not but we can easily rerun the tests.