@dkarrasch 2019-06-14
Could you quickly run the examples in #32303? Is it true that with this PR the `AbstractVector` specialization yields faster code than when artificially wrapping the vector in an array?