@chethega 2019-06-14
Before:
```
julia> begin
       haystack = repeat(["ab","cd","ef","gh"], 250);
       needles = ["ab","cd","gh"];
       @btime filter($(in(needles)), $haystack)
       @btime getindex($haystack, map($(in(needles)), $haystack))
       haystack = rand(100_000);
       pred = x->x<0.5
       @btime filter($pred, $haystack)
       @btime getindex($haystack, map($pred, $haystack))
       haystack = reshape(rand(100_000), 1,100000)
       @btime filter($pred, $haystack) 
       @btime getindex($haystack, map($pred, $haystack))
       ;
       end;
  26.094 μs (10 allocations: 16.39 KiB)
  20.837 μs (5 allocations: 7.14 KiB)
  1.219 ms (16 allocations: 1.00 MiB)
  684.314 μs (7 allocations: 488.91 KiB)
  691.666 μs (8 allocations: 488.52 KiB)
  690.217 μs (7 allocations: 488.47 KiB)
```
After:
```
  19.933 μs (1 allocation: 7.94 KiB)
  20.842 μs (5 allocations: 7.14 KiB)
  148.827 μs (3 allocations: 781.33 KiB)
  695.398 μs (7 allocations: 492.72 KiB)
  159.846 μs (3 allocations: 781.33 KiB)
  682.754 μs (7 allocations: 489.59 KiB)
```