@chethega 2019-06-15
> Is it true that with this PR the AbstractVector specialization yields faster code than when artificially wrapping the vector in an array?

To answer your question look at the benchmarks from the top:
```
p=0.5, N=1000000
filter!(pred, a::Vector)
  6.872 ms (0 allocations: 0 bytes) #before
  1.721 ms (1 allocation: 0 bytes) #after
filter!(pred, a::Test.GenericVector)
  6.962 ms (0 allocations: 0 bytes)
  1.674 ms (0 allocations: 0 bytes)
filter(pred, a::Vector)
  12.613 ms (19 allocations: 5.00 MiB)
  1.747 ms (3 allocations: 7.63 MiB)
filter(pred, a::Matrix)
  7.305 ms (9 allocations: 4.76 MiB)
  1.743 ms (3 allocations: 7.63 MiB)
filter(pred, a::Test.GenericVector)
  7.755 ms (9 allocations: 4.76 MiB)
  3.383 ms (6 allocations: 11.43 MiB)
filter(pred, a::AbstractMatrix) with linear indexing
  8.056 ms (9 allocations: 4.76 MiB)
  3.171 ms (6 allocations: 11.43 MiB)
filter(pred, a::AbstractMatrix) with cartesian indexing
  13.467 ms (9 allocations: 4.76 MiB)
  13.479 ms (7 allocations: 4.76 MiB)
```
You see that all `Array` now have the same performance, which is about 4x faster than the fastest old implementation and 8x faster than the old `Vector` implementation (when 0.5 of entries are filtered and the predicate is super cheap). 

AbstractArray with linear indexing (especially AbstractVector) uses a slower algorithm, because we cannot assume that its indices are `1:length(v)` (see discussion above) and I failed at figuring out a branch-free allocation-free version that uses the `eachindex` / `axes` protocol.