@dkarrasch 2019-06-15
This is amazing work! If I read the original benchmarks correctly, then this PR even improves the case when the success rate is very small? That's when the current `mapfilter` approach would `push!` very rarely and hence have little overhead.