@chethega 2019-06-15
Yes. The only case that should currently slow down is when you have a very cheap predicate and filter a very small (size k) subset of giant Array (size n) of non-isbits type.

In that case, the new algorithm allocates a new array of length n, fills it and afterwards truncates it to size k. That could be done almost free with appropriate syscalls, but is currently expensive (need to zero 8*n bytes twice, once in the kernel pagefault handler and once in the julia initializer for the array; optimally we would only write O(k) bytes to memory). The current mood appears to be against mucking with OS internals, but I hope that this will get fixed eventually.