@chethega 2019-06-17
Single test fail looks unrelated ("Error: Download failed: curl: (6) Could not resolve host: frippery.org").

Any change requests? Rebase & squash & merge? Or let it wait some more time?