@StefanKarpinski 2019-06-18
`frippery.org` is a super sketchy sounding domain to be downloading things from.